Loading of class name pool types; Use the type along with the name for ClassInfo loading (more robust); Set of component class; In supervisor stack traces correct the Java operation and address being swapped

+++ runt/klib/supervisor/cc/squirreljme/jvm/io/BinaryBlob.java
@@ -27,7 +27,7 @@ public abstract class BinaryBlob
 	 * @throws IndexOutOfBoundsException If the index is outside of bounds.
 	 * @since 2019/09/22
 	 */
-	public abstract int readByte(int __o);
+	public abstract byte readByte(int __o);
 	
 	/**
 	 * Returns the size of this blob.
